+++ src/app/components/main/feed/PostForm.js
@@ -3,17 +3,20 @@ import React from "react";
 
 const PostForm = (props) => {
     return (
-        < div id="modalPost" open={this.props.isVisible} className="modal" >
+        <form onSubmit={(e) =>{e.preventDefault()}}>
+        <div >
             <h3>{props.postName}</h3>
             <span>Post content</span>
             <div>
-                <textarea className="materialize-textarea" value={this.props.text} onChange={this.props.onChangeText}></textarea>
-                <button onClick={() => this.instances[0].close()}
-                    disabled={this.props.disabledText}
-                    className=" btn" >Post</button>
+                <input className="materialize-textarea" onCreateText={props.onCreateText}></input>
+                <button className=" btn" onPostText={props.onPostText}>Post</button>
             </div>
 
-        </div >
-
+        </div>
+</form>
     )
 }
+
+
+
+export default PostForm;
